feat: restore terminal app implementation
