+++ components/apps/terminal.worker.js
